Control.Evolution

Description

The evolution package supports the creation of evolutionary and bio-inspired algorithms with an user friendly domain specific language.

The package provides the main reproduction and selection strategies, a DSL to describe the evolution process, a function that runs the evolution algorithm with support to concurrency, and a customizable DSL to describe the Crossover and Mutation operators for specific class of problems.

The main idea is that the user focus only on implementing the crossover and mutation operators, the individual representation, a fitness function, and a random individual generator. After that, the user can customize the main loop of the evolutionary process using the DSL and without worrying about implementing the process.

Example - Simple GA:

Let's suppose we are working with a binary representation and we want to support a one-point crossover, single point mutation, multi-point mutation.

We must first create a data type with the chromossome and a fitness field and add support to `Eq, Ord` classes by comparing the fitness:

data Bin = Bin { _chromo :: [Bool]
               , _fit :: Double
               }
               
instance Eq Bin where
  t1 == t2 = _fit t1 == _fit t2
instance Ord Bin where
  t1 <= t2 = _fit t1 <= _fit t2

After that, we instantiate the Solution class to indicate how to retrieve the fitness of the individual and check whether it is feasible or not.

instance Solution Bin where
  _getFitness = _fit 
  _isFeasible = const 0.0 -- always feasible

Next, we instantiate the EvoClass to create the DSL for the crossover and mutation operators:

instance EvoClass Bin where 
  data Crossover Bin = OnePoint deriving (Show, Read)
  data Mutation  Bin = SinglePoint | MultiPoints deriving (Show, Read)

Finally, we implement the required functions and create an Interpreter:

cx OnePoint [p1,p2] = do
  let c1 = _chromo p1
      c2 = _chromo p2
  ix <- randomIndex (length c1)
  let c = take ix c1 <> drop ix c2
  pure $ Bin c 0.0 -- the individual will be evaluated later
  
mut SinglePoint (Bin c f) = do
  ix <- randomIndex (length c)
  let c' = swap ix c
  pure $ Bin c' 0.0
  
randomIndividual n = do
    c take n <$ state randoms
    pure $ Bin c 0.0
    
-- how many True in the chromossome
fitness (Bin c _) = pure $ Bin c (length $ filter id c)

interpret :: Interpreter Test
interpret = Funs cx mut randomIndividual fitness

We can create a customized logger to store the information at every generation.

logger :: Population a -> IO ()
logger pop  = do
  let best  = getBestFitness pop
      worst = getBestFitness worst
      avg   = getAvgFitness
  print $ "The best fitness is: " <> show best
  print $ "The worst fitness is: " <> show worst
  print $ "The avg fitness is: " <> show avg

Now you can define your evolutionary algorithm using the DSL and run it using runEvolution

evo = Reproduce Generational 
        [With AllOf 
           :> Cross OnePoint 2 0.3 (Tournament 2) 
           :> Mutate SinglePoint 0.7 
           :> Done
        ]

-- run the evolution with 100 individuals for 500 generations
-- and random seed g
runEvolution 500 100 logger evo g interpret

We can, for example, create a an evolutionary algorithm that creates two child population: one with crossover and mutation and another with only mutation, and then we reproduce using tournament selection.

evo = Reproduce (Probabilistic (Tournament 2))
        [ With AllOf :> Cross OnePoint 2 0.3 (Tournament 2) :> Mutate SinglePoint 0.7 :> Done
        , With AllOf :> Mutate MultiPoints 0.9 :> Done
        ]

We can even parse the evolution process from a file. If we have a file "ga.evo" with the content

Reproduce Generational 
    [With AllOf 
       :> Cross OnePoint 2 0.3 (Tournament 2) 
       :> Mutate SinglePoint 0.7 
       :> Done
    ]

We can automatically parse as:

evo read <$ readFile "ga.evo"

The evolution DSL is described as a reproduction operator followed by a list of evolution cycles.

Reproduce operator [cycle]

The evolution cycle can be either Parent, when you just want to pass the current population, `operator :> cycle`, when you want to chain two operators, and Done, marking the end of the cycle.

The operator can be

Cross operator n-parents probability operator
Mutate operator probability
With predicate

The With operator is used to filter the population and can be AllOf (no filter), Feasible, Infeasible.

Documentation

runEvolution

Arguments

:: (Solution a, NFData a)

=> Int
number of generations

-> Int
population size

-> (Population a -> IO ())
a logger function to run at every generation

-> Evolution a
Evolutionary process

-> StdGen
random number generator

-> Interpreter a
Interpreter of evolutionary process

-> IO ([Double], a)
returns the average fitness of every generation and the final champion

Runs the evolutionary process

type Rnd a = StateT StdGen IO a

A random element is the state monad with StdGen as a state

class (Show (Crossover t), Read (Crossover t), Show (Mutation t), Read (Mutation t)) => EvoClass t

Data family for Crossover and Mutation

as each algorithm/problem t has a different set of possible operators, the user must provide a data type for Crossover and Mutation.

See test/Specs.hs for an example.

Associated Types

data Crossover t :: *

data Mutation t :: *

type Population a = Vector a

A population of solutions is a vector

class Ord a => Solution a where

A Solution class must implement a function to evaluate this solution, a function to retrieve its fitness and another function to check the feasibility of the solution.

The evaluation function is inside the state monad as it may need to sample some instances to evaluate (i.e., lexicase selection).

Methods

_getFitness :: a -> Double

_isFeasible :: a -> Bool

data Interpreter a

The Interpreter data type stores all the necessary functions for the evolution process. Each of these functions should receive the type of operation as the first parameter and throw an error in case an operator is unsupported:

`_cx OnePoint pop = myOnePoint pop` `_cx NPoints pop = error "NPoints crossover not implemented for this algorithm"`

Constructors

Funs

Fields

_cx :: Crossover a -> [a] -> Rnd a

_mut :: Mutation a -> a -> Rnd a

_create :: Rnd a

_evaluate :: a -> Rnd a

data Reproduction

Reproduction algorithms

Generational: the children replaces the parent population

KeepBest: generational but keeping the best individual (either from parent or child population)

ReplaceWorst: replaces the worst individuals with new random solutions or sampled solutions from the parent

Merge: merge all the populations from different generation process

Probabilistic: sample individuals from the combined parent+child populations using a Selection strategy

Constructors

Generational

KeepBest

ReplaceWorst

Merge

Probabilistic Selection

data Selection

Selection algorithms

Tournament: tournament selection, samples n solutions and returns the best

RouletteWheel: roulette wheel, calculates the probability of choosing an individual proportional to its fitness and sample an individual using this probability distribution

Constructors

Tournament Int

RouletteWheel

data Predicate

Predicate for population filter

Feasible: selects only the feasible solutions

Infeasible: selects only the infeasible solutions

All: selects the entire population

Constructors

Feasible

Infeasible

AllOf

data Evolution t

DSL of an evolutionary process

The evolutionary process is composed of a reproduction strategy between two evolutionary cycles applied to a filtered population.

Constructors

Reproduce Reproduction [EvoCycle t]

data EvoCycle t

An evolution cycle is composed of a sequence of crossover and mutation operations. The order and quantity of each operation is arbitrary.

Cross: applies a Crossover operator to n individuals with p probability and sampling individuals using Selection strategy

Mutate: applies a Mutation operator with probability p

Common evolution cycle:

Cross OnePoint 2 0.8 (Tournament 2) (Mutate SwapTree 0.2 End)

With Local Search:

Cross OnePoint 2 0.8 (Tournament 2) (Mutate SwapTree 0.2 (Mutate LocalSearch 1.0 End))

data EvoCycle t = Loop (Gen t)

Constructors

(Op t) :> (EvoCycle t) infixr 5

Done

Parent

data Op t

Constructors

Cross (Crossover t) Int Double Selection

Mutate (Mutation t) Double

With Predicate